Partial revert of #197
This reverts the binrc update, the Hugo alias hack, but keeps the shared libraries to help people use Hugo extended, if they provide their own binary.

This is not ideal, but its the minimum assistance we can provide on the Ubuntu 14 image.  

To use Hugo extended on the Ubuntu 14 image, download the version of Hugo that you want, then run it like this:

```console
LD_LIBRARY_PATH=$HOME/stdc++6/usr/lib/x86_64-linux-gnu ./hugo
```