Added Netlify environment variable.
For downstream authors of linting/build tools that run in various CI
environments, it's important to be able to detect which environment the
build tool is being run in. This is important because of how Netlify
configures its environment, environment variables, dependencies, etc.
as they different from other CI environments like Travis CI or Circle
CI for example.

To illustrate further, when running on Circle CI or Travis CI, the
following environment variables are always present:

    CIRCLECI=true
    TRAVIS=true

The change being committed here adds Netlify support in the same manner
as found in other CI environments:

    NETLIFY=true

This provides consistency of behavior across the various CI
environments so authors, like me, can correctly detect which build
environment my tool is being run in and adjust accordingly.

This is of greatest benefit for tools like [Git
Cop](https://is.gd/UyxZGq) that needs to calculate commits on the
feature branch for linting purposes. If this pull request should be
accepted, I'll be able to update my
[documentation](https://github.com/bkuhlmann/git-cop#continuous-integration-ci)
accordingly to show I can support Netlify properly.