Pin bundler version so builds pinning it to old version keep working
Pinning bundler to 2.0.1 was recommended as a workaround to not being
able to install bundles created with the latest version of
bundler. That was caused by the build-image only using a single
version of bundler and then going un-updated for a long time. This
resulted in a lot of customers using Gemfile.locks with an old version
of bundler (2.0.1) listed as BUNDLED_WITH. When the build-image
underlying the buildbot was updated and deployed, all these builds
started failing because the build-image had the latest version of
bundler (2.0.2).

See https://github.com/netlify/build-image/issues/328 and
https://github.com/netlify/build-image/issues/331.

See also https://github.com/netlify/build-image/issues/296, which is
the long-term solution to this issue.