Cache clojure dependencies
This speeds up builds of Clojure repos by avoiding network calls to
download dependencies.

When a repo has a `deps.edn` file, we assume it is a Clojure repo and
would benefit from pre-populating and caching its primary dependencies.

Since Clojure dependencies are ultimately stored in the global `.m2`
directory, along with other Java dependencies, it suffices to restore
`.m2`. The directory is automatically populated during a build performed
with the [Clojure CLI tools][clojure-cli].

This extends 52f8eb9, which added support for repos which are built with
the Clojure CLI tools. It is similar to f3a4d49 which introduced caching
for Leiningen and Boot, two other ways of building Clojure code.

[clojure-cli]: https://clojure.org/reference/deps_and_cli