Cache Clojure classpath and prefill caches
For repos that are built with the [Clojure CLI tools][clojure-cli]
(i.e., repos marked with a `deps.edn` file), we cache and restore the
`.cpcache` directory. Clojure uses this directory to avoid unnecessarily
re-calculating the JVM classpath. For most projects this additional
cache will not lead to a large improvement in build times.

We also download dependencies to the `.m2` cache before handing off to
the build script. Technically, this is unnecessary since, presumably,
the build script uses the Clojure CLI tools which have a side-effect of
populating `.m2`. Therefore, the benefit of this change is simply to
separate the two steps, isolating any errors that occure while fetching
dependencies.

[clojure-cli]: https://clojure.org/reference/deps_and_cli