Replace xenial with focal in README.md
Following the steps in the `README.md` to the letter leads to downloading the docker image twice. Once for xenial (which is no longer supported by Netlify builds) and once for focal (in Step 2, once `./test-tools/start-image.sh path/to/site/repo` is executed).

This commit removes mention of the `netlify/build:focal` image to align this with step two.